refactor(orchestrator): extract compression-guideline learning into CompressionGuidelineCoordinator (#1526)
Seam 4 of the orchestrator decomposition (#1526). Moves the
compression-guideline learning lifecycle (candidate computation,
LLM semantic refinement, draft persistence + activation, the
recall section, and the consolidation-time learning pass) out of
the 20.9k-LOC orchestrator into a new dependency-injected
orchestration/compression-guideline-coordinator.ts.

Behavior-preserving: the orchestrator keeps thin delegating
methods so the public API (optimizeCompressionGuidelines,
activateCompressionGuidelineDraft) and the two internal hooks
(consolidation loop + recall pipeline) are unchanged.
formatCompressionGuidelinesForRecall moves with it and is
re-exported to preserve the public surface.

orchestrator.ts: 20989 -> 20650 LOC (-339).
entity-hardening: 246/246 green; orchestrator-characterization
snapshots unchanged.

Chokepoints used: none (pure extraction, no new surfaces).
